MetaPod first source push.
This is a working version of MetaPod that is able to load 32x portable executables and inject custom metadata into them. The library is setup to build as a C shared lib.

Further work is needed on error handling, cleaning up the API, documentation, and making a basic wrapper in C#.